# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/dunning_runs:
    get:
      tags:
        - DunningRuns
      summary: List DunningRuns
      description: Get a paginated list of DunningRuns with optional filtering and sorting
      operationId: listDunningRuns
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DunningRunList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - DunningRuns
      summary: Create DunningRun
      description: Create a new DunningRun
      operationId: createDunningRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDunningRunRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DunningRun'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/dunning_actions:
    get:
      tags:
        - DunningActions
      summary: List DunningActions
      description: Get a paginated list of DunningActions with optional filtering and sorting
      operationId: listDunningActions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DunningActionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - DunningActions
      summary: Create DunningAction
      description: Create a new DunningAction
      operationId: createDunningAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDunningActionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DunningAction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    DunningRun:
      type: object
      required:
        - id
        - company_id
        - as_of_date
        - direction
        - actions_emitted
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        as_of_date:
          type: string
          format: date
        direction:
          type: string
          maxLength: 10
        snapshot_id:
          nullable: true
          type: string
          format: uuid
        actions_emitted:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    CreateDunningRunRequest:
      type: object
      required:
        - company_id
        - as_of_date
        - direction
        - actions_emitted
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        as_of_date:
          type: string
          format: date
        direction:
          type: string
          maxLength: 10
        snapshot_id:
          nullable: true
          type: string
          format: uuid
        actions_emitted:
          type: integer
          format: int32
        status:
          type: string
        metadata:
          type: object
...
    DunningRunList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DunningRun'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    DunningAction:
      type: object
      required:
        - id
        - company_id
        - run_id
        - invoice_ref
        - invoice_kind
        - level
        - action_type
        - days_past_due
        - outstanding_amount
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
          maxLength: 10
        party_id:
          nullable: true
          type: string
          format: uuid
        level:
          type: string
        action_type:
          type: string
        days_past_due:
          type: integer
          format: int32
        outstanding_amount:
          type: string
        status:
          type: string
        processed_at:
          nullable: true
          type: string
          format: date-time
        result_ref:
          nullable: true
          type: string
          maxLength: 140
        metadata:
          type: object
...
    CreateDunningActionRequest:
      type: object
      required:
        - company_id
        - run_id
        - invoice_ref
        - invoice_kind
        - level
        - action_type
        - days_past_due
        - outstanding_amount
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        invoice_ref:
          type: string
          format: uuid
        invoice_kind:
          type: string
          maxLength: 10
        party_id:
          nullable: true
          type: string
          format: uuid
        level:
          type: string
        action_type:
          type: string
        days_past_due:
          type: integer
          format: int32
        outstanding_amount:
          type: string
        status:
          type: string
        processed_at:
          nullable: true
          type: string
          format: date-time
        result_ref:
          nullable: true
          type: string
          maxLength: 140
        metadata:
          type: object
...
    DunningActionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DunningAction'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []